requirement.txt
numpy
onnx
onnxruntime-gpu
opencv-python-headless